# util/sets: simplify List() by using slices.Sort