--- a/staging/src/k8s.io/apimachinery/pkg/util/sets/set_generic_test.go
+++ b/staging/src/k8s.io/apimachinery/pkg/util/sets/set_generic_test.go
@@ -356,3 +356,20 @@ func TestIntersection(t *testing.T) {
 		}
 	}
 }
+
+func BenchmarkListSmall(b *testing.B) {
+	s := sets.New("a", "b", "c", "d", "e", "f", "g", "h", "i", "j")
+	for b.Loop() {
+		sets.List(s)
+	}
+}
+
+func BenchmarkListLarge(b *testing.B) {
+	s := make(sets.Set[int], 12345)
+	for i := range 12345 {
+		s.Insert(i * 7)
+	}
+	for b.Loop() {
+		sets.List(s)
+	}
+}