@huww98 2024-11-17
With go 1.23, we can simplify this even further. Now List() and UnsortedList() only have one line of code each! PTAL.