@huww98 2024-11-18
It seems the new implementation based on go 1.23 iterator is returning nil for empty set, and does not match the current behavior. So let me revert to the previous version.

I also added a benchmark, which shows the new implementation is 4x faster:
```
go test -benchmem -run=^$ -bench ^BenchmarkList k8s.io/apimachinery/pkg/util/sets
```
master branch:
```
goos: darwin
goarch: arm64
pkg: k8s.io/apimachinery/pkg/util/sets
cpu: Apple M2 Pro
BenchmarkListSmall-12            1766280               841.6 ns/op           184 B/op          2 allocs/op
BenchmarkListLarge-12                662           2454698 ns/op          106520 B/op          2 allocs/op
```

This PR **(BEST, 4x faster)**:
```
goos: darwin
goarch: arm64
pkg: k8s.io/apimachinery/pkg/util/sets
cpu: Apple M2 Pro
BenchmarkListSmall-12            4719645               236.5 ns/op           160 B/op          1 allocs/op
BenchmarkListLarge-12               1939            608282 ns/op          106496 B/op          1 allocs/op
```

Also tested the (abandoned) iterator implementation:
```
goos: darwin
goarch: arm64
pkg: k8s.io/apimachinery/pkg/util/sets
cpu: Apple M2 Pro
BenchmarkListSmall-12            1000000              1054 ns/op             560 B/op          8 allocs/op
BenchmarkListLarge-12                786           1316229 ns/op          488769 B/op         23 allocs/op
```
The performance is pool, and the allocation is high